+++ CHANGELOG.md
@@ -1,2 +1,13 @@
+#### 1.0.1 (2020-01-14)
+- **2DAffine**: Add parabolic saturation at the output as the shearing may increase a lot the signal dynamic.
+
+- **MarokSeq**: 
+	1. Add output triggers for each state
+	2. Add tools to set probabilities related to a given state to 0.0 or to 0.5
+	3. Modify the initialization such that the sequencer moves as a regular sequencer (forward direction)
+	4. Add a context menu to choose the behavior if all probabilities are 0
+
+ 
+
 #### 1.0.0 (2020-01-08)
 - Initial Release of 2DAffine, 2DRotation, MarkovSeq, PolginalVCO.